+++ pkgs/google_cloud_storage/CHANGELOG.md
@@ -1,7 +1,8 @@
 ## 0.6.2-wip
 
 * Add retries to `uploadObjectFromSink`.
-* Add `deleteObjectAcl`.
+* Add `deleteObjectAcl`, `getObjectAcl`, `listObjectAcl`, `patchObjectAcl`,
+  and `updateObjectAcl`.
 * Add `makeObjectPublic`.
 
 ## 0.6.1